pyproject.toml
[project]
name = "seo-audit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "a2a-sdk==0.2.4",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bs4>=0.0.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "deprecated==1.2.18",
    "fastapi==0.115.12",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "json-repair==0.47.8",
    "lxml>=6.0.0",
    "opentelemetry-api==1.33.1",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.0",
    "requests==2.32.4",
    "sniffio==1.3.1",
    "sse-starlette==2.3.5",
    "starlette==0.46.2",
    "tldextract>=5.3.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "wrapt==1.17.2",
    "zipp==3.22.0",
]
